Update README.md
Added: "Christensen also stated that disruption is partly about consumers establishing the purpose of technologies through practice (Christensen and Raynor, 2003).

Markides’ (2006) added on to this and stated that disruptive technologies do not happen naturally, but are created through practice. Michael Flavin, Head of Curriculum Innovation at King's College London, adds on to Markides’ argument and cited Wikipedia as an example: encyclopedias are established forms of publication, but Wikipedia has changed the way they are distributed, produced, and consumed. 

Therefore, astute practice of a technology gives rise to disruption and not its intrinsic features (Christensen and Raynor, 2003). This is because practice “does not occur in a vacuum but is shaped by underpinning economic and social factors (Flavin, 2012).” " under WHAT ARE DISRUPTIVE TECHNOLOGIES?

Added: "This also further strengthens Flavin's (2012) aforementioned statement about how economic and social factors influence the rise of disruptive technologies." under HOW DOES ONLINE LEARNING DISRUPT TRADITIONAL FACE-TO-FACE EDUCATION? 

Added: Sources